# app2/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Ticketing System API
  version: 1.0.0
servers:
  - description: Ticketing System API
    url: '{scheme}://{address}:{port}'
    variables:
      address:
        default: localhost
      port:
        default: "8081"
      scheme:
        default: http
        enum:
          - https
          - http
paths:
...
  /features:
    get:
      summary: List all features
      operationId: listFeatures
      parameters:
        - { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "default": 0 } }
        - { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10 } }
      responses:
        "200": { "description": "A list of features", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Feature" } } } } }
    post:
      summary: Create a new feature
      operationId: createFeature
      requestBody: { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewFeature" } } } }
      responses:
        "200": { "description": "Features created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Feature" } } } }
...
components:
  schemas:
...
    NewFeature:
      type: object
      properties:
        name: { "type": "string" }
      required: [ "name" ]
    Feature:
      type: object
      properties:
        id: { "type": "string" }
        name: { "type": "string" }
        created: { "type": "string" }
        updated: { "type": "string" }
      required: [ "id", "name", "created", "updated" ]
...